backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
google-generativeai==0.8.3
playwright==1.49.1
httpx==0.28.1
structlog==24.4.0
pytest==8.3.4
pytest-asyncio==0.25.0
google-genai>=0.8.0
